requirements.txt
# Core dependencies
anthropic>=0.18.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
chromadb>=0.4.22
sentence-transformers>=2.3.0

# Data processing
pydantic>=2.5.0
python-dotenv>=1.0.0
pyyaml>=6.0

# API and web
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0

# Utilities
rich>=13.7.0
tqdm>=4.66.0
numpy>=1.26.0
pandas>=2.1.0
